Add samples test file for config_file
